+++ src/duk_initjs.js
@@ -18,11 +18,6 @@
         def(D, name, value);
     }
 
-    // Logger object for C code provided by init code now.
-    if (true) {
-        def(D.Logger, 'clog', new D.Logger('C'));
-    }
-
     // Tracking table for CommonJS module loading.
     if (true) {
         def(D, 'modLoaded', {});